Split off XML and JSON functions
