+++ golang/monogram.go
@@ -161,146 +161,6 @@ func parseToAST(input string) []*Node {
 	return []*Node{root}
 }
 
-func printASTXML(nodes []*Node, indentDelta string, output io.Writer) {
-	for _, node := range nodes {
-		printNodeXML(node, "", indentDelta, output)
-	}
-}
-
-func printNodeXML(node *Node, currentIndent string, indentDelta string, output io.Writer) {
-	// Open the XML tag
-	fmt.Fprintf(output, "%s<%s", currentIndent, node.Name)
-
-	// Print attributes (with escaping)
-	for key, value := range node.Options {
-		fmt.Fprintf(output, ` %s="%s"`, key, escapeXMLValue(value))
-	}
-
-	// Handle self-closing tag if no children are present
-	if len(node.Children) == 0 {
-		fmt.Fprintln(output, " />")
-		return
-	}
-
-	// Otherwise, close the opening tag and iterate over children
-	fmt.Fprintln(output, ">")
-	newIndent := currentIndent + indentDelta
-	for _, child := range node.Children {
-		printNodeXML(child, newIndent, indentDelta, output)
-	}
-
-	// Close the XML tag
-	fmt.Fprintf(output, "%s</%s>\n", currentIndent, node.Name)
-}
-
-func escapeXMLValue(value string) string {
-	replacer := strings.NewReplacer(
-		"&", "&amp;",
-		"<", "&lt;",
-		">", "&gt;",
-		"\"", "&quot;",
-		"'", "&apos;",
-	)
-	return replacer.Replace(value)
-}
-
-func printASTJSON(nodes []*Node, indentDelta string, output io.Writer) {
-	// Start the currentIndent with the indentDelta
-	currentIndent := indentDelta
-
-	// Open the JSON array
-	fmt.Fprintf(output, "[\n")
-	for i, node := range nodes {
-		printNodeJSON(node, currentIndent, indentDelta, output) // Adjust child indentation
-		if i < len(nodes)-1 {
-			fmt.Fprintln(output, ",") // Add a comma for all but the last node
-		} else {
-			fmt.Fprintln(output)
-		}
-	}
-	// Close the JSON array
-	fmt.Fprintf(output, "]\n")
-}
-
-func printNodeJSON(node *Node, currentIndent string, indentDelta string, output io.Writer) {
-	// Precompute the next level of indentation
-	nextIndent := currentIndent + indentDelta
-
-	// Open the object
-	fmt.Fprintf(output, "%s{\n", currentIndent)
-
-	// Include the `role` field without escaping the name
-	fmt.Fprintf(output, "%s\"role\": \"%s\",\n", nextIndent, node.Name)
-
-	// Flatten the options map directly into string-valued fields
-	optionCount := len(node.Options)
-	current := 0
-	for key, value := range node.Options {
-		current++
-		escapedValue := escapeJSONString(value)              // Escape the option value
-		if current < optionCount || len(node.Children) > 0 { // Add a comma if there are more fields or children
-			fmt.Fprintf(output, "%s\"%s\": \"%s\",\n", nextIndent, key, escapedValue)
-		} else {
-			fmt.Fprintf(output, "%s\"%s\": \"%s\"\n", nextIndent, key, escapedValue)
-		}
-	}
-
-	// Print the children field
-	if len(node.Children) > 0 {
-		fmt.Fprintf(output, "%s\"children\": [\n", nextIndent)
-
-		childIndent := nextIndent + indentDelta
-		for i, child := range node.Children {
-			printNodeJSON(child, childIndent, indentDelta, output)
-			if i < len(node.Children)-1 {
-				fmt.Fprintln(output, ",") // Add a comma for all but the last child
-			} else {
-				fmt.Fprintln(output)
-			}
-		}
-
-		fmt.Fprintf(output, "%s]\n", nextIndent) // Close the JSON array for children
-	}
-
-	// Close the object
-	fmt.Fprintf(output, "%s}", currentIndent)
-}
-
-func escapeJSONString(value string) string {
-	var sb strings.Builder
-
-	for _, r := range value {
-		switch r {
-		// JSON escape sequences for special characters
-		case '"':
-			sb.WriteString("\\\"") // Escape double quotes
-		case '\\':
-			sb.WriteString("\\\\") // Escape backslashes
-		case '\b':
-			sb.WriteString("\\b") // Escape backspace
-		case '\f':
-			sb.WriteString("\\f") // Escape form feed
-		case '\n':
-			sb.WriteString("\\n") // Escape newline
-		case '\r':
-			sb.WriteString("\\r") // Escape carriage return
-		case '\t':
-			sb.WriteString("\\t") // Escape tab
-
-		// Printable ASCII characters (no escaping required)
-		default:
-			if r >= 0x20 && r <= 0x7E {
-				sb.WriteRune(r)
-			} else {
-				// For non-ASCII and non-printable characters, use \u escape
-				sb.WriteString(fmt.Sprintf("\\u%04X", r))
-			}
-		}
-	}
-
-	return sb.String()
-}
-
 func translate(input io.Reader, output io.Writer, printAST func([]*Node, string, io.Writer), indentSpaces int) {
 	// Read the entire input as a string
 	data, err := io.ReadAll(input)
